+++ composeApp/src/commonMain/kotlin/com/inspiredandroid/kai/ui/settings/SettingsUiState.kt
@@ -58,7 +58,6 @@ enum class SettingsTab {
     Services,
     Tools,
     Sandbox,
-    Integrations,
 }
 
 @Immutable